Refactor inheritance
